% by naoi okada
% charset=utf-8
\encoding{utf-8}
\name{package.install}
\alias{package.install}
\title{
[Package] Install Packages which are not already Installed
}
\description{
リポジトリにないパッケージやバイナリがないパッケージも含めてインストールする。
}
\usage{
package.install (\dots)
}
\arguments{
\item{\dots}{ロードするパッケージ名のベクトル。}
}
